pyproject.toml
[project]
name = "my-final-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "frontend>=0.0.3",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "pymupdf>=1.26.7",
    "PyPDF2>=3.0.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "fsspec>=2025.3.2",
    "google-genai>=1.58.0",
    "ipython>=8.36.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=0.3.16",
    "langchain-text-splitters>=0.3.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "langchain-qdrant>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "huggingface-hub>=0.36.0",
    "langchain-elasticsearch>=1.0.0",
    "elasticsearch>=8.19.3",
    "streamlit>=1.53.0",
    "ragas>=0.4.3",
    "datasets>=4.5.0",
    "pandas>=2.3.3",
]
